auth-service/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
httpx
python-dotenv